Changed float to uint8

+++ outfacingInterfaces/Helpers/Color.hpp
@@ -9,14 +9,14 @@ class Color {
 public:
     Color() = default;
 
-    Color(float r, float g, float b, float a) : r(r), g(g), b(b), a(a) {};
+    Color(uint8_t r, uint8_t g, uint8_t b, uint8_t a) : r(r), g(g), b(b), a(a) {};
 
     ~Color() = default;
 
-    float r;
-    float g;
-    float b;
-    float a;
+    uint8_t r;
+    uint8_t g;
+    uint8_t b;
+    uint8_t a;
 };
 
 #endif //BRACK_ENGINE_COLOR_HPP
